+++ protocol/src/utils/UniV3Abstraction.sol
@@ -85,7 +85,7 @@ library UniV3Abstraction {
     }
 
     function swap(
-        uint160 accountId,
+        uint256 accountId,
         address pool,
         bool amountIsInput,
         bool isVEthToVGas,
